+++ frontend/src/pages/notice/NoticeWritePage/index.tsx
@@ -2,8 +2,8 @@ import { Notice } from '@type/notice';
 
 import { useCreateNotice } from '@hooks';
 
-import AdminNoticeWrite from '@components/notice/AdminNoticeWrite';
-import { NOTICE_WRITE_TYPE } from '@components/notice/AdminNoticeWrite/constant';
+import NoticeForm from '@components/notice/NoticeForm';
+import { NOTICE_WRITE_TYPE } from '@components/notice/NoticeForm/constant';
 
 import { addTimeToDate } from '@utils/post/addTimeToDate';
 
@@ -21,10 +21,6 @@ export default function NoticeWritePage() {
   const { mutate } = useCreateNotice();
 
   return (
-    <AdminNoticeWrite
-      notice={defaultNotice}
-      writeType={NOTICE_WRITE_TYPE.CREATE}
-      writeNotice={mutate}
-    />
+    <NoticeForm notice={defaultNotice} writeType={NOTICE_WRITE_TYPE.CREATE} writeNotice={mutate} />
   );
 }
